+++ TS51huapao/TSmyzhixiaoTableViewController.m
@@ -75,7 +75,7 @@
 //    if (_priceRange != nil) {
 //        [dic setObject:_priceRange forKey:@"priceRange"];
 //    }
-    [dic setObject:[NSString stringWithFormat:@"%u",_page] forKey:@"pageindex"];
+    [dic setObject:[NSString stringWithFormat:@"%lu",(unsigned long)_page] forKey:@"pageindex"];
     
     NSURLSessionDataTask * task = [TSGetItemListPost globalTimeGetRecommendInfoWithDictionary:dic Block:^(NSArray *posts,NSUInteger maxcount, NSError *error) {
         if (!error) {
